working prototype of board generator
with better printing out, yay
